Fix chat input text box not showing up.
The string wasn't found in the string table because doesStringExist does
a case-sensitive search for the string and it starts with an uppercase
character, but the string in the string table is all lowercase. Since it
was not found, KeyPress defaults to using the char value of the keysym,
which is '\r'.